moved chart components into separate folder
