+++ client/Routes.js
@@ -6,14 +6,14 @@ import Signup from './components/Signup';
 import Questionnaire from './components/Questionnaire';
 import Dashboard from './components/Dashboard';
 import { me } from './store';
-import Budget from './components/BudgetChart';
+import Budget from './components/ChartComponents/BudgetChart';
 import AllTransactions from './components/AllTransactions';
 import AddTransactions from './components/AddTransactions';
 import EditTransactions from './components/EditTransactions';
 import EditSingleTransaction from './components/EditSingleTransaction';
 import UserProfile from './components/UserProfile';
-import BudgetHistoryLineGraph from './components/BudgetHistoryLineGraph';
-import GraphContainer from './components/GraphContainer';
+import BudgetHistoryLineGraph from './components/ChartComponents/BudgetHistoryLineGraph';
+import GraphContainer from './components/ChartComponents/GraphContainer';
 
 /**
  * COMPONENT
