+++ client/components/ChartComponents/BarGraphCategory.js
